feat!: bump nrf sdk support from v2.6.2 to v2.9.1 (#56)
The following commit includes changes from multiple PRs, and are included the following project changes:
* Bump nrf sdk support from v2.6.1 to 2.7.0 and finally to v2.9.1
* Force sysbuild as github action build type
* Update readme to match the new hardware model
* Update GitHub actions docker image to use the official nrf sdk image
* Update readme
* Move CONFIG_SOFTSIM_BUNDLE_TEMPLATE_HEX to sysbuild.
* Fix compiler warnings when printing eDRX parameters.
* Move PM_STATIC configuration to sysbuild.
* Add static partition layout for Thingy:91 X.

A big shout out and kudos to Stig Bjørlykke <stig.bjorlykke@nordicsemi.no> at Nordic Semiconductor as the primary contributor to this transition to Zephyr Sysbuild.
And a big thank you to Mirko Covizzi <mirko.covizzi@nordicsemi.no> as well for guidance during this effort.

---------

+++ samples/softsim_static_profile/src/main.c
@@ -73,8 +73,8 @@ static void lte_handler(const struct lte_lc_evt *const evt)
       char log_buf[60];
       ssize_t len;
 
-      len = snprintf(log_buf, sizeof(log_buf), "eDRX parameter update: eDRX: %f, PTW: %f", evt->edrx_cfg.edrx,
-                     evt->edrx_cfg.ptw);
+      len = snprintf(log_buf, sizeof(log_buf), "eDRX parameter update: eDRX: %f, PTW: %f",
+                     (double)evt->edrx_cfg.edrx, (double)evt->edrx_cfg.ptw);
       if (len > 0) {
         LOG_INF("%s\n", log_buf);
       }
